Fix Word class' handling of unknown words
The Word class uses natlink.getWordInfo() to retrieve formatting
information. That function returns None if the word is not present in
the dictionary. Before this change, that resulted in an error. After
this change, such unknown words are treated as if they don't have any
formatting info.

This change also adds test cases for the Word class.