Fix backslash processing in dictation results
This change causes the last backslash character to be used instead of
the first to split written and spoken forms in dictation results for
DNS v10. Using the first breaks on the backslash character
("\\backslash") where the written form itself is a backslash.

This change also adds some minor improvements to the word
formatting doctests.