Fix bug adding lists with same content
This change fixes a bug that prevented multiple lists from being added
to a grammar if their content was the same. A check meant to prevent
a single list from being added multiple times incorrectly checked for
identical content instead of identity.

This change adds new doctests showing how to use list constructs and
verifying that the bug fixed in this change does not occur.

Resolves #36.