Make Sphinx engine's module loader check for local engine config
The module loader will now use a local engine 'config.py' file in
its directory if it exists and has all of the attributes set.

Warnings are printed if there is a config.py file without the
engine config attributes set. The loader will fallback on the
default engine config.py module in that case.