Add ContextAction class from Aenea
ContextAction is a class from Aenea that executes different actions
depending on which context is currently active. This version of the
class works without Aenea's functionality.

I have left the original license & copyright notice at the top of the
file.

This commit also adds some documentation, error handling for executing
actions/matching contexts, as well as a usage example.