Add optional 'remap_data' argument to the Function action
The new argument allows using extras with different names than the
function argument names. A usage example/doctest has been added to
the action's documentation. This is compatible with the 'defaults'
argument.

This commit also fixes some print statements in the Function action
doctests and changes a relative import so that the doctests run
correctly.