Use the correct dx and dy returned from `EnumDisplayMonitors`
The [EnumDisplayMonitors](https://docs.microsoft.com/en-us/windows/desktop/api/winuser/nf-winuser-enumdisplaymonitors) function returns a [RECT structure](https://docs.microsoft.com/en-us/previous-versions//dd162897(v=vs.85)). The fields in this struct refer to the x,y coordinates of the top left and bottom right corners of a monitor, not the origin and dx,dy as the original code assumed.

This change uses the correct calculations for the Monitor object to behave correctly.