Add optional 'repeating' parameter to the multiplexing Timer class
This parameter is True by default. If it is False, then the timer's
function will only be called once. The stop() method will be called
afterwards to prevent further calls.

This commit also adds the parameter to the engine.create_timer()
methods as well as a unit test in the test_timer.py file.