Add threading fix and related method to the NatlinkEngine class
This changes the Natlink engine class to start a thread and engine
timer internally to allow Python threads to work properly. The fix
is applied automatically when connect() is called or when a grammar
is loaded for the first time. The fix can also be applied manually
by calling the NatlinkEngine.apply_threading_fix() method.