Use Dragonfly's Win32 Clipboard class on Windows
Dragonfly was changed to use a cross-platform clipboard class a
while ago. This simply changes the default Clipboard class to the
original Win32 Clipboard class on Windows, as it supports more
clipboard formats than the cross-platform class.

This change is backwards-compatible; both classes have the same
methods and the cross-platform class was based on the Win32 class.