Change the X11Window.maximize() method to use the wmctrl program
The wmctrl program can adjust the Extended Window Manager Hints
(EWMH) for window maximization in most X11 window managers. This
is better than using a keyboard shortcut.

It should be noted that neither the current xdotool or xprop
release versions can maximize the window by EWMH, so the new
external dependency is necessary.