Change internal integer classes to support path modification
Re: #306.

This allows language implementations of the Integer class to modify
and add new elements for words in recognition paths associated with
IntBuilders.

This functionality is useful in circumstances where one or more
words in a builder spec should be affixed in some way to possible
remainder, multiplier and/or element word(s).