Adjust internal integer path modification behaviour
This changes the relevant classes to have two modes of operation
for recognition path modification: AUGMENT and REPLACE.

The first mode is the default behaviour, which notably skips path
strings unchanged by the modifier function. It only uses paths that
*augment* the original element tree. The second mode keeps and uses
unchanged path strings, since the collection element is to *replace*
the original element tree.