Change how Win32Clipboard handles irretrievable formats
The GetClipboardData() function sometimes raises errors for valid
clipboard formats. Specifically, this occurs if the Shell format
CFSTR_FILECONTENTS is on the Windows system clipboard when the
copy_from_system() method is called. Other formats may be affected.

To fix this, such errors are now caught and logged if the problem
format was not specified using the copy_from_system() *formats*
argument.