Adjust order of preferred SR engine implementations in get_engine()
Re: #330.

The Kaldi engine back-end is now preferred and used over the SAPI 5
and Sphinx back-ends, if it is available.  The reasoning for this
change is that Kaldi is generally more accurate, more efficient and
available on more platforms than the other two.

This change only makes a difference when get_engine() is not given
a specific engine name.  The SAPI 5 and Sphinx back-ends are still
initialized if requested or if the Kaldi back-end is not available.