Improve Windows keyboard action support for letter and number keys
This changes the Windows keyboard code to fallback on predefined
virtual-key codes for keys in ranges 0-9, a-z and A-Z if the key
cannot be typed using the current keyboard layout.  This means that
the equivalent key will be typed instead, allowing common keyboard
shortcuts, such as Ctrl+C, to work properly.