Change window mover classes to use threads instead of engine timers
Engine timers are used by these classes to visibly move windows from
one position to another.  This would have been done to allow the
functionality to be used with the Natlink/DNS engine back-end, which
then had multi-threading limits.

These limits have not been an issue since PR #232 was merged, so
this can now be done with threads instead.  This change also nicely
detaches the functionality from the engine code.