Clean up some code added in PR #340 and fix regression
The regression was that letter and number characters would not be
consistently typed with the Unicode keyboard interface if hardware
events were not required and the current keyboard layout did not
support typing the characters.  The Key action might press the
equivalent keys on the layout using hardware events, while the Text
action would use Unicode events appropriately.