Add missing logger names and default levels into log.py
The "dictation.word_parser" and "dictation.word_parser_factory"
loggers are now set to level 30 (WARNING) by default, for the
stderr and file handlers.  These loggers are used in the dictation
formatting code for the Natlink and DNS engine.

These changes only apply if the setup_log() function is used.