Change win32_clipboard_ctx() to handle nested calls properly
I've changed the function into a context manager class instead and
have made it so the Windows CloseClipboard() function is called only
once in the __exit__() method of the first and outermost manager
instance.

I've used a thread-local counter since it seems that Windows
manages clipboard access on per thread basis.