Fix incorrect handling of extended scan codes on Windows
This changes the KeyboardInput struct in sendinput.py to use the
MAPVK_VK_TO_VSC_EX map type when translating virtual key codes on
modern versions of Windows.  The KEYEVENTF_EXTENDEDKEY flag is added
only if the scan code has a high byte of 0xe0 or 0xe1.

This removes the hardcoded list of extended keys.