Fix a potential bug with mouse button emulation on Windows
Re: #185.

If an error occurs when emulating mouse button events with the
SendInput() function and the primary mouse button was set by the
user to the right button, then that setting would not be restored,
leaving our left-handed user understandably confused the next time
they use the mouse.  This fixes that potential bug by swapping the
primary button back in a 'finally' block instead.