Remove English-language alternatives for the integer two
Re: #84, #85.

The alternatives here removed are "to" and "too".  These were added
to fix a bug that occasionally occurs with the DNS engine back-end.
In hindsight, it seems unnecessary for all back-ends to include this
confusing behaviour.  Use of custom integer content would seem to be
a better solution.

Test cases using these alternatives have been removed or adjusted as
necessary.