Fix command-line interface bug with non-matching patterns
Fixes #357.

This fixes a bug where non-matching glob patterns are not treated
as normal filenames.  For example, if the pattern "_*.py" matches no
files in the working directory, then the *load* and *test* commands
will attempt to use the pattern string as a filename and give the
usual error message if that fails.