Set Kaldi engine option auto_add_to_user_lexicon to False by default
This option normally causes the engine to query the website
*www.speech.cs.cmu.edu* for the pronunciation of each unknown word
encountered during the grammar loading process.

It is now turned off by default because, since it sends user data to
a remote server, it should be opt-in.  It is also possible for the
remote connection to fail in some circumstances, which causes the
engine to become unresponsive.