Make SR engine text-to-speech functionality more flexible
The (limited) text-to-speech (TTS) functionality of the Natlink and
SAPI 5 engine back-ends has been moved into separate classes.  This
makes it possible to, for instance, use the SAPI 5 TTS functionality
independently of the speech recognition part.

The *engine.speak()* methods of the Kaldi, CMU Pocket Sphinx and the
text-input SR engine back-ends will use the first available TTS imp-
lementation out of "sapi5", "natlink" and "text".  The first two are
only available on Windows. The third will simply write the specified
text to *stdout*.

A *get_speaker()* function has been added.  It works very much like
the *get_engine()* function, returning an instance of the default or
specified *Speaker* class.