Migrated to PEP517/pyproject.toml compatible package.  Project dependancies, version number, metadata etc now in pyproject.toml.
However, the build step to produce sphix documentation is not supported since setup.py is not used for this packaging.  setuptools integration is scheduled to be removed from setuptools in V7.0.
This is the largest change.

The code to run tests has beem moved to main_test.py, and pyproject.toml has test configuration so pytest in the project root runs the test.

changes to .travis.yml have not been tested.

It will not be possible to build the package with python 2.7; however it should be possible to install the package with pip.  The mechanism to run the tests will be different for python 2.7; Also the test dependencies are different
for python 2.7.  `pip install .[python27_test]` to bring in the correct version of pytest.

The build system is flit (replacing setuptools).  An editiable install  'pip install -e .' from the project root should bring in the build tool requirements and pytest, or 'pip install flit'.