Add feature allowing key and mouse events to be sent via Natlink
Re: #372.

Natlink's `playEvents' function can be used to simulate keystrokes
and mouse events with Dragon NaturallySpeaking.  Dragonfly's Key,
Text and Mouse action objects are now able to send events this way,
allowing Dragonfly commands to interact with administrative applic-
ations on Windows.

This feature is disabled by default because the `playEvents' inter-
face is not a perfect substitute for the Windows `SendInput' funct-
ion.  Dragonfly falls back on the default implementation if Natlink
cannot be used.

Documentation on this feature has been added in the relevant files.