Add aliases for typing the Windows modifier key
Capital C and capital S may now be used as prefix characters in Key
action spec strings to specify Windows key keystrokes.  This means,
for example, that "C-space", "S-space" and "w-space" are equivalent
spec strings.  C is for Command and S is for Super; the names of
the Windows key on MacOS and X11 respectively.

Equivalents to the "win" and "rwin" typeables have also been added.

For simplicity, these typeables have been made available on each
supported platform.