Add aliases "Option" and "Meta" for typing the Alt modifier key
Capital O and capital M may now be used as prefix characters in Key
action spec strings to specify Alt key keystrokes.  O is for Option
and M is for Meta; the names sometimes used for this key on MacOS
and X11 respectively.

As with the new Windows key aliases, equivalent typeables have also
been added.