Change the behaviour of Repetition and Integer max constructor arguments
Re: #394.

The *max* constructor arguments of the Repetition and Integer class-
es are now treated as inclusive, rather than exclusive.  This seems
more sensible and matches the treatment of the *min* argument.

For the Integer class, I have elected not to change the complex int-
ernal classes that do the most of the work.  They receive the value
*max + 1*.  This shouldn't matter, but I'm noting it here.

The Digits element class is affected by this change, as it is a sub-
class of Repetition.